+++ CHANGELOG.md
@@ -37,4 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1](https://github.com/DFE-Digital/polis-whitelabel/issues/1) Fixed `authorUid is undefined` crash in `server`
 - [#3](https://github.com/DFE-Digital/polis-whitelabel/issues/3) Fixed crash when `.headersJson` file not present for `file-server`
 
+### Performance
+
+- [#80](https://github.com/DFE-Digital/polis-whitelabel/issues/80) Reduce client-participation bundlesize by optimizing D3.
+
 ## Forked from [github.com/compdemocracy/polis/commit/52e983a](https://github.com/compdemocracy/polis/commit/52e983aef272c140a77a5b26ea6bf0ead9f8aece)